+++ src/simmate/apps/chemspace/datastores/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .freedom_ro5 import FreedomRo5MoleculeStore